refactored test-package location
