+++ bgw-gui/src/test/kotlin/tools/aqua/bgw/components/uicomponents/structureddataview/StructuredDataViewTestBase.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package tools.aqua.bgw.uicomponents.structureddataview
+package tools.aqua.bgw.components.uicomponents.structureddataview
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
