+++ bgw-gui/src/test/kotlin/tools/aqua/bgw/components/uicomponents/structureddataview/listview/ListViewChangeSelectionModeTest.kt
@@ -15,16 +15,16 @@
  * limitations under the License.
  */
 
-package tools.aqua.bgw.uicomponents.structureddataview.tableview
+package tools.aqua.bgw.components.uicomponents.structureddataview.listview
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import tools.aqua.bgw.components.uicomponents.ListView
 import tools.aqua.bgw.components.uicomponents.SelectionMode
-import tools.aqua.bgw.components.uicomponents.TableView
 
-/** Test changing [SelectionMode] in [TableView]. */
-class TableViewChangeSelectionModeTest : TableViewTestBase(SelectionMode.SINGLE) {
+/** Test changing [SelectionMode] in [ListView]. */
+class ListViewChangeSelectionModeTest : ListViewTestBase(SelectionMode.SINGLE) {
 
   /** Test change to [SelectionMode.NONE]. */
   @Test
