+++ bgw-gui/src/test/kotlin/tools/aqua/bgw/components/uicomponents/structureddataview/listview/ListViewTestSelectionModeSingle.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package tools.aqua.bgw.uicomponents.structureddataview.listview
+package tools.aqua.bgw.components.uicomponents.structureddataview.listview
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
@@ -24,8 +24,8 @@ import org.junit.jupiter.api.assertThrows
 import tools.aqua.bgw.components.uicomponents.ListView
 import tools.aqua.bgw.components.uicomponents.SelectionMode
 
-/** Tests for [ListView] with [SelectionMode.MULTIPLE]. */
-class ListViewTestSelectionModeMultiple : ListViewTestBase(SelectionMode.MULTIPLE) {
+/** Tests for [ListView] with [SelectionMode.SINGLE]. */
+class ListViewTestSelectionModeSingle : ListViewTestBase(SelectionMode.SINGLE) {
 
   /** Test select by index on valid parameter. */
   @Test
@@ -147,7 +147,6 @@ class ListViewTestSelectionModeMultiple : ListViewTestBase(SelectionMode.MULTIPL
     dataView.items.clear()
 
     assertThrows<IllegalArgumentException> { dataView.selectFirst() }
-
     checkNotNotified()
     checkInvocation(0, 0, 0)
   }
@@ -180,28 +179,92 @@ class ListViewTestSelectionModeMultiple : ListViewTestBase(SelectionMode.MULTIPL
   @Test
   @DisplayName("Test select next")
   fun testSelectNext() {
-    assertThrows<IllegalStateException> { dataView.selectNext() }
-    checkNotNotified()
-    checkInvocation(0, 0, 0)
+    dataView.selectNext()
+
+    checkNotified()
+    checkInvocation(1, 0, 0)
+    assertEquals(0, invokedIndex) // None was selected, so index 0 should be selected now
+
+    // Inject selected index to selection model
+    dataView.selectedIndicesList.add(0)
+
+    // Second invocation
+    dataView.selectNext()
+
+    checkNotified(2)
+    checkInvocation(2, 0, 0)
+    assertEquals(1, invokedIndex) // Next index, i.e. 1 should be selected now
+  }
+
+  /** Test select next if last is selected. */
+  @Test
+  @DisplayName("Test select next if last is selected")
+  fun testSelectNextLastSelected() {
+    dataView.selectLast()
+    checkNotified()
+    checkInvocation(1, 0, 0)
+    assertEquals(items.size - 1, invokedIndex) // Last should be selected
+
+    // Inject selected index to selection model
+    dataView.selectedIndicesList.add(items.size - 1)
+
+    // Second invocation
+    dataView.selectNext()
+
+    checkNotified(1) // Should not notify
+    checkInvocation(1, 0, 0) // Should not invoke
   }
 
   /** Test select previous. */
   @Test
   @DisplayName("Test select previous")
   fun testSelectPrevious() {
-    assertThrows<IllegalStateException> { dataView.selectPrevious() }
-    checkNotNotified()
-    checkInvocation(0, 0, 0)
+    dataView.selectPrevious()
+
+    checkNotified()
+    checkInvocation(1, 0, 0)
+    assertEquals(
+        items.size - 1, invokedIndex) // None was selected, so last index should be selected now
+
+    // Inject selected index to selection model
+    dataView.selectedIndicesList.add(items.size - 1)
+
+    // Second invocation
+    dataView.selectPrevious()
+
+    checkNotified(2)
+    checkInvocation(2, 0, 0)
+    assertEquals(
+        items.size - 2,
+        invokedIndex) // Previous index, i.e. (items.size - 1) -1 should be selected now
+  }
+
+  /** Test select previous if last is selected. */
+  @Test
+  @DisplayName("Test select previous if last is selected")
+  fun testSelectPreviousFirstSelected() {
+    dataView.selectFirst()
+    checkNotified()
+    checkInvocation(1, 0, 0)
+    assertEquals(0, invokedIndex) // First should be selected
+
+    // Inject selected index to selection model
+    dataView.selectedIndicesList.add(0)
+
+    // Second invocation
+    dataView.selectPrevious()
+
+    checkNotified(1) // Should not notify
+    checkInvocation(1, 0, 0) // Should not invoke
   }
 
   /** Test select all. */
   @Test
   @DisplayName("Test select all")
   fun testSelectAll() {
-    dataView.selectAll()
-
+    assertThrows<IllegalStateException> { dataView.selectAll() }
     checkNotNotified()
-    checkInvocation(0, 1, 0)
+    checkInvocation(0, 0, 0)
   }
 
   /** Test select all when items are empty. */
@@ -209,17 +272,15 @@ class ListViewTestSelectionModeMultiple : ListViewTestBase(SelectionMode.MULTIPL
   @DisplayName("Test select all when items are empty")
   fun testSelectAllEmptyList() {
     dataView.items.clear()
-
-    assertThrows<IllegalArgumentException> { dataView.selectAll() }
-
+    assertThrows<IllegalStateException> { dataView.selectAll() }
     checkNotNotified()
     checkInvocation(0, 0, 0)
   }
 
-  /** Test clear selection one selected. */
+  /** Test clear selection. */
   @Test
-  @DisplayName("")
-  fun testClearSelectionOneSelected() {
+  @DisplayName("Test clear selection")
+  fun testClearSelection() {
     dataView.select(2)
     checkNotified()
     checkInvocation(1, 0, 0)
@@ -232,23 +293,6 @@ class ListViewTestSelectionModeMultiple : ListViewTestBase(SelectionMode.MULTIPL
 
     checkNotified(2)
     checkInvocation(1, 0, 1) // Clear should have been invoked
-  }
-
-  /** Test clear selection all selected. */
-  @Test
-  @DisplayName("Test clear selection all selected")
-  fun testClearSelectionAllSelected() {
-    dataView.selectAll()
-    checkNotified()
-    checkInvocation(0, 1, 0)
-
-    // Inject selected index to selection model
-    dataView.selectedIndicesList.addAll(items.indices.toList())
-
-    dataView.clearSelection()
-
-    checkNotified(2)
-    checkInvocation(0, 1, 1) // Clear should have been invoked
   }
 
   /** Test clear selection if none was selected. */
