+++ bgw-gui/src/test/kotlin/tools/aqua/bgw/components/uicomponents/structureddataview/tableview/TableViewTestSelectionModeNone.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package tools.aqua.bgw.uicomponents.structureddataview.tableview
+package tools.aqua.bgw.components.uicomponents.structureddataview.tableview
 
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
