+++ bgw-gui/src/test/kotlin/tools/aqua/bgw/observable/observablelist/SingleListFunctionTests.kt
@@ -132,7 +132,6 @@ class SingleListFunctionTests : ObservableListTestBase() {
   @DisplayName("Test sort")
   fun testSort() {
     list.sort(Comparator.comparingInt { it })
-
     for (i in list.indices) assertEquals(ordered[i], list[i])
 
     checkNotified()
